+++ app/src/main/java/io/itch/fr/quizgame/MainActivity.kt
@@ -3,44 +3,26 @@ package io.itch.fr.quizgame
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
+import androidx.activity.viewModels
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import io.itch.fr.quizgame.ui.theme.QuizGameTheme
+import io.itch.fr.quizgame.feature_start.presentation.StartPage
+import io.itch.fr.quizgame.feature_start.presentation.StartViewModel
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            QuizGameTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    Greeting("Android")
-                }
-            }
+            TriviaQuizApp()
         }
     }
-}
 
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
+    @Composable
+    private fun TriviaQuizApp() {
+        val startViewModel: StartViewModel by viewModels()
 
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    QuizGameTheme {
-        Greeting("Android")
+        StartPage(
+            onStartQuizClicked = { /* Handle start quiz button click */ },
+            onViewHistoryClicked = { /* Handle view history button click */ }
+        )
     }
 }