+++ app/src/main/java/io/itch/fr/quizgame/feature_start/presentation/StartViewModel.kt
@@ -1,0 +1,7 @@
+package io.itch.fr.quizgame.feature_start.presentation
+
+import androidx.lifecycle.ViewModel
+
+class StartViewModel : ViewModel() {
+    // Add any necessary logic or data for the Start page here
+}